Organized code and added length attribute to car

+++ src/TrafficLight.java
@@ -3,9 +3,9 @@ import org.json.JSONObject;
 public class TrafficLight {
     private JSONObject JSONTrafficLight = new JSONObject();
     private char curr_color;
-    private float curr_duration;
+    private double curr_duration;
 
-    TrafficLight(char color, float curr_duration) {
+    TrafficLight(char color, double curr_duration) {
         this.curr_color = color;
         this.curr_duration = curr_duration;
